Finished up Available downloads view
